Fixed Free, again?.
"We can remove the header as it's not needed" Yeah thanks past James, real cool

Added a current settings printout for the settings. Not sure if I like it - perhaps have it merge with the index list?

This would make it easier to read but would have to be printed each time